Revert "refactor: platform.io"
