--- .gitignore
@@ -1,7 +1,0 @@
-.pio
-.vscode/.browse.c_cpp.db*
-.vscode/c_cpp_properties.json
-.vscode/launch.json
-.vscode/ipch
-docs/JC2432W328
-include/emu_mac_address.h
