--- platformio.ini
@@ -1,29 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env:esp32dev]
-platform = espressif32
-board = esp32dev
-framework = arduino
-lib_deps = 
-	bodmer/TFT_eSPI@^2.5.43
-	lvgl/lvgl@8.4.0
-	mbed-seeed/BluetoothSerial@0.0.0+sha.f56002898ee8
-build_flags=  
-	; configure lvlgl
-	-DLV_CONF_INCLUDE_SIMPLE
-	; this folder has the lv_conf.h. Add it to the global include path
-	-Ilib/
-	; configure TFT_eSPI
-	-D USER_SETUP_LOADED=1
-	; -include $PROJECT_LIBDEPS_DIR/$PIOENV/TFT_eSPI/User_Setups/Setup1_ILI9341.h
-	-include $PROJECT_DIR/lib/User_Setup.h
-board_build.partitions = partitions_singleapp_3MB.csv
-monitor_speed = 115200
